requirements.txt
python-telegram-bot>=20.0
quart>=0.18.0
pandas
numpy
requests
scikit-learn
tensorflow
prophet
ta
textblob
ccxt
nest_asyncio